Use flutter_media_metadata instead of metadata_god

+++ lib/data/audio.dart
@@ -25,7 +25,7 @@ class Audio {
   final String? title;
 
   /// The ID3-Tag duration of the audio file. It can be null if was not set
-  final double? durationMs;
+  final int? durationMs;
 
   /// The ID3-Tag artist(s) of the audio file.
   final String? artist;
@@ -95,7 +95,7 @@ class Audio {
     String? description,
     String? website,
     String? title,
-    double? durationMs,
+    int? durationMs,
     String? artist,
     String? album,
     String? albumArtist,
